+++ src/constants/look.rs
@@ -26,6 +26,12 @@ use crate::{
     traits::{IntoExpectedType, TryInto},
 };
 
+#[cfg(feature = "score")]
+use crate::objects::{ScoreCollector, ScoreContainer};
+
+#[cfg(feature = "symbols")]
+use crate::objects::{SymbolContainer, SymbolDecoder};
+
 /// Internal enum representing each LOOK_* constant.
 ///
 /// It's recommended to use the constants in the `look` module instead for type
@@ -40,7 +46,7 @@ use crate::{
 /// See the [module-level documentation][crate::constants] for more details.
 #[doc(hidden)]
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash, Serialize_repr, Deserialize_repr, FromStr)]
-#[repr(u8)]
+#[repr(u16)]
 pub enum Look {
     #[display("creep")]
     Creeps = 0,
@@ -72,6 +78,18 @@ pub enum Look {
     Deposits = 12,
     #[display("ruin")]
     Ruins = 13,
+    #[cfg(feature = "score")]
+    #[display("scoreContainer")]
+    ScoreContainers = 10011,
+    #[cfg(feature = "score")]
+    #[display("scoreCollector")]
+    ScoreCollectors = 10012,
+    #[cfg(feature = "symbols")]
+    #[display("symbolContainer")]
+    SymbolContainers = 10021,
+    #[cfg(feature = "symbols")]
+    #[display("symbolDecoder")]
+    SymbolDecoders = 10022,
 }
 
 js_deserializable!(Look);
@@ -111,4 +129,16 @@ typesafe_look_constants! {
     pub struct TOMBSTONES = (Look::Tombstones, Tombstone, IntoExpectedType::into_expected_type);
     pub struct POWER_CREEPS = (Look::PowerCreeps, PowerCreep, IntoExpectedType::into_expected_type);
     pub struct RUINS = (Look::Ruins, Ruin, IntoExpectedType::into_expected_type);
+}
+
+#[cfg(feature = "score")]
+typesafe_look_constants! {
+    pub struct SCORE_CONTAINERS = (Look::ScoreContainers, ScoreContainer, IntoExpectedType::into_expected_type);
+    pub struct SCORE_COLLECTORS = (Look::ScoreCollectors, ScoreCollector, IntoExpectedType::into_expected_type);
+}
+
+#[cfg(feature = "symbols")]
+typesafe_look_constants! {
+    pub struct SYMBOL_CONTAINERS = (Look::SymbolContainers, SymbolContainer, IntoExpectedType::into_expected_type);
+    pub struct SYMBOL_DECODERS = (Look::SymbolDecoders, SymbolDecoder, IntoExpectedType::into_expected_type);
 }