+++ src/objects.rs
@@ -122,6 +122,22 @@ reference_wrappers! {
     pub struct AccountPowerCreep(...);
 }
 
+#[cfg(feature = "score")]
+reference_wrappers! {
+    #[reference(instance_of = "ScoreContainer")]
+    pub struct ScoreContainer(...);
+    #[reference(instance_of = "ScoreCollector")]
+    pub struct ScoreCollector(...);
+}
+
+#[cfg(feature = "symbols")]
+reference_wrappers! {
+    #[reference(instance_of = "SymbolContainer")]
+    pub struct SymbolContainer(...);
+    #[reference(instance_of = "SymbolDecoder")]
+    pub struct SymbolDecoder(...);
+}
+
 /// Trait for things which have positions in the Screeps world.
 ///
 /// This can be freely implemented for anything with a way to get a position.
@@ -211,6 +227,18 @@ impl_has_id! {
     StructureWall,
     Tombstone,
     PowerCreep,
+}
+
+#[cfg(feature = "score")]
+impl_has_id! {
+    ScoreContainer,
+    ScoreCollector,
+}
+
+#[cfg(feature = "symbols")]
+impl_has_id! {
+    SymbolContainer,
+    SymbolDecoder,
 }
 
 /// Trait for all wrappers over Screeps JavaScript objects extending
@@ -491,6 +519,10 @@ unsafe impl Transferable for StructureTower {}
 unsafe impl Transferable for StructurePowerSpawn {}
 unsafe impl Transferable for StructureTerminal {}
 unsafe impl Transferable for PowerCreep {}
+#[cfg(feature = "score")]
+unsafe impl Transferable for ScoreCollector {}
+#[cfg(feature = "symbols")]
+unsafe impl Transferable for SymbolDecoder {}
 
 // NOTE: keep impls for Structure* in sync with accessor methods in
 // src/objects/structure.rs
@@ -507,6 +539,10 @@ unsafe impl Withdrawable for StructureTower {}
 unsafe impl Withdrawable for StructurePowerSpawn {}
 unsafe impl Withdrawable for StructureTerminal {}
 unsafe impl Withdrawable for Tombstone {}
+#[cfg(feature = "score")]
+unsafe impl Withdrawable for ScoreContainer {}
+#[cfg(feature = "symbols")]
+unsafe impl Withdrawable for SymbolContainer {}
 
 unsafe impl Harvestable for Deposit {}
 unsafe impl Harvestable for Mineral {}
@@ -573,6 +609,14 @@ unsafe impl RoomObjectProperties for StructureWall {}
 unsafe impl RoomObjectProperties for Structure {}
 unsafe impl RoomObjectProperties for Tombstone {}
 unsafe impl RoomObjectProperties for PowerCreep {}
+#[cfg(feature = "score")]
+unsafe impl RoomObjectProperties for ScoreContainer {}
+#[cfg(feature = "score")]
+unsafe impl RoomObjectProperties for ScoreCollector {}
+#[cfg(feature = "symbols")]
+unsafe impl RoomObjectProperties for SymbolContainer {}
+#[cfg(feature = "symbols")]
+unsafe impl RoomObjectProperties for SymbolDecoder {}
 
 impl_structure_properties! {
     OwnedStructure,
@@ -624,6 +668,7 @@ unsafe impl OwnedStructureProperties for StructureTower {}
 
 unsafe impl HasStore for Creep {}
 unsafe impl HasStore for Ruin {}
+
 unsafe impl HasStore for StructureContainer {}
 unsafe impl HasStore for StructureExtension {}
 unsafe impl HasStore for StructureFactory {}
@@ -637,6 +682,12 @@ unsafe impl HasStore for StructureTerminal {}
 unsafe impl HasStore for StructureTower {}
 unsafe impl HasStore for Tombstone {}
 unsafe impl HasStore for PowerCreep {}
+#[cfg(feature = "score")]
+unsafe impl HasStore for ScoreContainer {}
+#[cfg(feature = "score")]
+unsafe impl HasStore for ScoreCollector {}
+#[cfg(feature = "symbols")]
+unsafe impl HasStore for SymbolContainer {}
 
 // NOTE: keep impls for Structure* in sync with accessor methods in
 // src/objects/structure.rs
@@ -666,3 +717,7 @@ unsafe impl CanDecay for StructurePortal {}
 unsafe impl CanDecay for StructureRampart {}
 unsafe impl CanDecay for StructureRoad {}
 unsafe impl CanDecay for Tombstone {}
+#[cfg(feature = "score")]
+unsafe impl CanDecay for ScoreContainer {}
+#[cfg(feature = "symbols")]
+unsafe impl CanDecay for SymbolContainer {}