+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Unreleased
 - Fixed `Room::serialize_path` and `Room::deserialize_path`, which are static methods and don't
   exist on instances of `Room` objects themselves.
 - Changed `BuildEvent` to match what's returned by the game, which doesn't match the documentation.
+- Add the `generate-pixel`, `inter-shard-memory`, and `score` features which enable features not
+  present in all server environments.
+- Add the `mmo` feature which activates the `generate-pixel` and `inter-shard-memory` because
+  these interfaces don't exist on private servers.
+- Add the `seasonal-season-1` feature for season 1, which activates the `score` feature.
+- Add the `symbols` feature to support season 2.
+- Add the `seasonal-season-2` feature for season 2, which activates the `symbols` feature.
 
 0.9.0 (2021-01-23)
 ==================